Fix file-size ratchet: split report-run builder into a FundStructureEndpoints partial
D1d's TryBuildReportRunAuditTrail pushed FundStructureEndpoints.cs to 2990 lines,
over its 2969 baseline cap (the ratchet runs before the build, hence verify-fast's
"no test summary"). Make FundStructureEndpoints a partial class and move the
report-run builder into FundStructureEndpoints.ReportingRunStream.cs — the same
split pattern as WorkstationEndpoints.Stream.cs, and the natural home for the D2 SSE
endpoint. Main file is back at the cap; the new partial still reaches the private
ProjectReportingRunAuditTrail. (Three unrelated files remain over cap from pre-existing
main growth — out of scope here.)
